+++ lib/src/cmaf/ham/types/mapper/hls/Byterange.ts
@@ -1,0 +1,1 @@
+export type Byterange = { length: number; offset: number };