+++ lib/src/cmsd/CMSD_STATIC.ts
@@ -5,4 +5,4 @@
  *
  * @beta
  */
-export const CMSD_STATIC = 'CMSD-Static';
+export const CMSD_STATIC: string = 'CMSD-Static';