+++ lib/src/cta/608/VerboseLevel.ts
@@ -1,0 +1,14 @@
+/**
+ * Logging levels for the CTA-608 parser.
+ *
+ * @group CTA-608
+ * @beta
+ */
+export const enum VerboseLevel {
+	ERROR = 0,
+	TEXT = 1,
+	WARNING = 2,
+	INFO = 2,
+	DEBUG = 3,
+	DATA = 3
+}