+++ lib/src/cta/utils/isValid.ts
@@ -9,10 +9,10 @@ import { CmValue } from '../CmValue.js';
  *
  * @internal
  */
-export const isValid = (value: CmValue) => {
+export function isValid(value: CmValue): boolean {
 	if (typeof value === 'number') {
 		return Number.isFinite(value);
 	}
 
 	return value != null && value !== '' && value !== false;
-};
+}