+++ lib/src/structuredfield/serialize/serializeDecimal.ts
@@ -36,7 +36,7 @@ import { serializeError } from './serializeError.js';
 //      digits) to output.
 //
 // 10.  Return output.
-export function serializeDecimal(value: number) {
+export function serializeDecimal(value: number): string {
 	const roundedValue = roundToEven(value, 3); // round to 3 decimal places
 	if (Math.floor(Math.abs(roundedValue)).toString().length > 12) {
 		throw serializeError(value, DECIMAL);