+++ lib/src/structuredfield/utils/INTEGER_DECIMAL.ts
@@ -1,4 +1,4 @@
 import { DECIMAL } from './DECIMAL.js';
 import { INTEGER } from './INTEGER.js';
 
-export const INTEGER_DECIMAL = `${INTEGER} or ${DECIMAL}`;
+export const INTEGER_DECIMAL: string = `${INTEGER} or ${DECIMAL}`;